Add files via upload
Business Goal :
To predict whether the Loan will be provided by the borrower/bank or not. It will be used by the management to understand how exactly the Loan Status vary with the independent variables like Gender, Married, Dependents, Education, Self_Employed Applicant Income etc. 
Further, the model will be a good way for management to decide their credit policy for existing and new customers.